requirements.txt
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.3
platformdirs==3.8.0
psycopg2==2.9.6
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
six==1.16.0
SQLAlchemy==2.0.17
SQLAlchemy-Utils==0.41.1
typing_extensions==4.7.0
urllib3==2.0.3
Werkzeug==2.3.6
zeep==4.2.1
